requirements.txt
plaid-python>=14.0.0
requests>=2.31.0
slack-sdk>=3.23.0
notion-client>=2.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytz>=2024.1
